Class type Protocol_client_context.full

The class type full allows to create contexts that are explicitly used by low-level shell functions, while containing various information (I/O services, RPCs...). Then, depending on the usage, the type may be coerced into one of its following ascendants to serve for explicit operations on blocks, chain or daemon for instance.

inherit Tezos_client_base.Client_context.full

The class Client_context.full provides I/O services for the client, the wallet, etc.

inherit (Tezos_shell_services.Shell_services.chain * Tezos_shell_services.Shell_services.block) Tezos_protocol_017_PtNairob.Environment.RPC_context.simple

Base interface provided to call RPCs, i.e., communication with the node. A client context is defined by mapping all RPCs protocol-generic to a specific protocol.

inherit [Tezos_shell_services.Shell_services.chain, Tezos_shell_services.Shell_services.block] Tezos_protocol_017_PtNairob.Environment.proto_rpc_context

Protocol RPCs exposed through the environment (using an additional chainpath).