Module Tezos_layer2_utils_017_PtNairob.Layer1_services

type Tezos_base.TzPervasives.error +=

| Cannot_read_block_metadata of Tezos_base.TzPervasives.Block_hash.t

type 'accu successful_operation_processor = {

apply : 'kind. 'accu -> source:Tezos_protocol_017_PtNairob.Protocol.Alpha_context.public_key_hash -> 'kind Tezos_protocol_017_PtNairob.Protocol.Alpha_context.manager_operation -> 'kind Tezos_protocol_017_PtNairob.Protocol.Apply_results.successful_manager_operation_result -> 'accu;

apply_internal : 'kind. 'accu -> source:Tezos_protocol_017_PtNairob.Protocol.Alpha_context.public_key_hash -> 'kind Tezos_protocol_017_PtNairob.Protocol.Apply_internal_results.internal_operation -> 'kind Tezos_protocol_017_PtNairob.Protocol.Apply_internal_results.successful_internal_operation_result -> 'accu;

}

type 'accu operation_processor = {

apply : 'kind. 'accu -> source:Tezos_protocol_017_PtNairob.Protocol.Alpha_context.public_key_hash -> 'kind Tezos_protocol_017_PtNairob.Protocol.Alpha_context.manager_operation -> 'kind Tezos_protocol_017_PtNairob.Protocol.Apply_results.manager_operation_result -> 'accu;

apply_internal : 'kind. 'accu -> source:Tezos_protocol_017_PtNairob.Protocol.Alpha_context.public_key_hash -> 'kind Tezos_protocol_017_PtNairob.Protocol.Apply_internal_results.internal_operation -> 'kind Tezos_protocol_017_PtNairob.Protocol.Apply_internal_results.internal_operation_result -> 'accu;

}

val process_manager_operations : 'a -> Tezos_client_017_PtNairob.Protocol_client_context.Alpha_block_services.operation list list -> 'a operation_processor -> 'a

process_manager_operations accu operations operator folds over the list of manager operations in operations applying operator to transform accu along the way.

val process_applied_manager_operations : 'a -> Tezos_client_017_PtNairob.Protocol_client_context.Alpha_block_services.operation list list -> 'a successful_operation_processor -> 'a

process_applied_manager_operations accu operations operator folds over the list of applied manager operations in operations applying operator to transform accu along the way.