Module type Context_helpers.P

module Tree : Tezos_context_sigs.Context.TREE with type key = string list and type value = bytes

Tree provides immutable, in-memory partial mirror of the context, with lazy reads and delayed writes. The trees are Merkle trees that carry the same hash as the part of the context they mirror.

type tree = Tree.tree

val hash_tree : tree -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.State_hash.t

type proof

val proof_encoding : proof Tezos_base.TzPervasives.Data_encoding.t

val proof_before : proof -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.State_hash.t

val proof_after : proof -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.State_hash.t

val verify_proof : proof -> (tree -> (tree * 'a) Lwt.t) -> (tree * 'a) option Lwt.t

val produce_proof : Tree.t -> tree -> (tree -> (tree * 'a) Lwt.t) -> (proof * 'a) option Lwt.t