Module Tezos_smart_rollup_017_PtNairob.Game_helpers

val default_new_dissection : default_number_of_sections:int -> start_chunk: Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Game.dissection_chunk -> our_stop_chunk: Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Game.dissection_chunk -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Tick.t list

default_new_dissection ~default_number_of_sections ~start_chunk ~our_stop_chunk computes a list of intermediary ticks that can later be turned into new dissection from between start_chunk and our_stop_chunk with make_dissection. The algorithm satisfies the default predicate on dissection exported by the protocol.

val make_dissection : state_of_tick: (?start_state:'a -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Tick.t -> ('a option, 'trace) Stdlib.result Lwt.t) -> state_hash_of_eval_state: ('a -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.State_hash.t) -> ?start_state:'a -> start_chunk: Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Dissection_chunk.t -> our_stop_chunk: Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Dissection_chunk.t -> Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Tick.t list -> (Tezos_protocol_017_PtNairob.Protocol.Alpha_context.Sc_rollup.Dissection_chunk.t list, 'trace) Stdlib.result Lwt.t

make_dissection ~state_hash_from_tick ~start_chunk ~our_stop_chunk intermediary_ticks computes a new dissection from a list of intermediary ticks between start_chunk and our_stop_chunk.

This function assumes intermediary_ticks encodes a valid dissection from start_chunk to our_stop_chunk, and recomputes the state hash associated to each ticks.

module Wasm : sig ... end